requirements.txt
jinja2>=3.0.1
PyYAML>=5.4.1
numpy>=1.20.2
bokeh>=2.3.3
aiohttp>=3.7.4
elasticsearch>=7.13.3
pandas>=1.3.2
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64

requests>=2.25.1
GoogleNews>=1.6.6
python-dateutil>=2.8.2